reverting r4463:4462

+++ jetty/branches/jetty-7-jaspi/modules/server/jetty/src/main/java/org/mortbay/jetty/AbstractConnector.java
@@ -29,7 +29,7 @@ import org.mortbay.thread.ThreadPool;
 
 
 /** Abstract Connector implementation.
- * This abstract implementation of the Connector interface provides:<ul>
+ * This abstract implemenation of the Connector interface provides:<ul>
  * <li>AbstractLifeCycle implementation</li>
  * <li>Implementations for connector getters and setters</li>
  * <li>Buffer management</li>
@@ -39,6 +39,8 @@ import org.mortbay.thread.ThreadPool;
  * </ul>
  * 
  * @author gregw
+ *
+ * TODO - allow multiple Acceptor threads
  */
 public abstract class AbstractConnector extends AbstractBuffers implements Connector
 {
@@ -219,7 +221,7 @@ public abstract class AbstractConnector extends AbstractBuffers implements Conne
     /**
      * @return Returns the soLingerTime.
      */
-    public int getSoLingerTime()
+    public long getSoLingerTime()
     {
         return _soLingerTime;
     }
